refactor: centralize booster card selection
